requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2==2.9.9
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==8.0.2
httpx==0.27.0